+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+* Issue **#419** : Fix multiple xml processing instructions appearing in output.
+
 * Issue **#446** : Fix for deadlock on rolling appenders.
 
 ## [v5.0-beta.50] - 2017-10-13